Add ScoreCounterCfgManager.
Simplify SettingsViewModel.

Add _isPersisted field to AppCfgManager and ScoreCounterCfgManager to hold state of Persist button (if it should be enabled) even when the screen is out of composition.

Load Score Counter config only when connection to Score Counter event occurred, not every time when navigating to Score Counter configuration tab.