Load SC config conditionally also when Settings screen is composed.
In theoretical case that Score Counter is connected sooner than SCCMListener from ScoreCounterCfgManager is registered, the SC config will not be loaded. Therefore, check if it is loaded when the SC Settings screen is composed and if not, load it.

Get rid of ScoreCounterCfgManager dependency in ScoreCounterConnectionManager, use callback instead.